+++ src/opcode.h
@@ -7,6 +7,8 @@
 #ifndef OPCODE_H
 #define OPCODE_H
 
+#include <stdint.h>
+
 #define MAXARG_Bx        (0xffff)
 #define MAXARG_sBx       (MAXARG_Bx>>1)         /* `sBx' is signed */
 
@@ -14,13 +16,13 @@
 /*              OP:A:Bx  = 7:9:16            */
 /*              OP:Ax    = 7:25              */
 
-#define GET_OPCODE(i) ((int)(((mrb_code)(i)) & 0x7f))
-#define GETARG_A(i)   ((int)((((mrb_code)(i)) >> 23) & 0x1ff))
-#define GETARG_B(i)   ((int)((((mrb_code)(i)) >> 14) & 0x1ff))
-#define GETARG_C(i)   ((int)((((mrb_code)(i)) >>  7) & 0x7f))
-#define GETARG_Bx(i)  ((int)((((mrb_code)(i)) >>  7) & 0xffff))
-#define GETARG_sBx(i) ((int)(GETARG_Bx(i)-MAXARG_sBx))
-#define GETARG_Ax(i)  ((int)((((mrb_code)(i)) >>  7) & 0x1ffffff))
+#define GET_OPCODE(i) ((int_fast8_t)(((mrb_code)(i)) & 0x7f))
+#define GETARG_A(i)   ((int_fast16_t)((((mrb_code)(i)) >> 23) & 0x1ff))
+#define GETARG_B(i)   ((int_fast16_t)((((mrb_code)(i)) >> 14) & 0x1ff))
+#define GETARG_C(i)   ((int_fast8_t)((((mrb_code)(i)) >>  7) & 0x7f))
+#define GETARG_Bx(i)  ((int_fast16_t)((((mrb_code)(i)) >>  7) & 0xffff))
+#define GETARG_sBx(i) ((int_fast16_t)(GETARG_Bx(i)-MAXARG_sBx))
+#define GETARG_Ax(i)  ((int_fast32_t)((((mrb_code)(i)) >>  7) & 0x1ffffff))
 #define GETARG_UNPACK_b(i,n1,n2) ((int)((((mrb_code)(i)) >> (7+n2)) & (((1<<n1)-1))))
 #define GETARG_UNPACK_c(i,n1,n2) ((int)((((mrb_code)(i)) >> 7) & (((1<<n2)-1))))
 #define GETARG_b(i)   GETARG_UNPACK_b(i,14,2)
